Political Affliation of LegCo Members

A list of data frames containing the political affiliation and constituency of each Legco members in each term.

legco_member_affiliation

Format

A list containing 1 dataframe with 10 variables. Each data frame represents one ternm, with the `TermID` as the name of the dataframe.

SpeakerID

The id of the member used in the Hansard Database

MemberID

The id of the member used in the Meeting Schedule Database

NameEng

English full name of the member

NameChi

Chinese full name of the member

TypeEng

The type of consituency that the member belongs to in English

TypeChi

The type of consituency that the member belongs to in Chinese

ConstituencyEng

The consituency that the member belongs to in English

ConstituencyChi

The consituency that the member belongs to in Chinese

AffiliationEng

The declared political affiliation of the member in English

AffiliationChi

The declared political affiliation of the member in Chinese

Details

Political affiliation is what the member has declared to LegCo, not necessary a full list of real-life party membership.

Examples

# NOT RUN {
legco_member_affiliation[["5"]]
# }